Bird Species Flask App/requirements.txt
Flask==2.2.5
keras==3.4.0
numpy==2.1.2
Pillow==11.0.0
